Refactor `api.js` (#81)
* Change `api.js` path from `/services` folder to `/services/api` folder
* Use API_URL in `api.js`
