+++ front-end/services/api/api.js
@@ -1,5 +1,5 @@
 export default async function fetchDevLinks() {
-  const url = 'https://us-central1-devlinktag-dev.cloudfunctions.net/api/devlink/all'; // TODO : 임시로 dev DB 사용
+  const url = `${process.env.API_URL}/devlink/all`;
   const response = await fetch(url);
   const data = await response.json();
   return data;
