+++ front-end/src/common/slice.js
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import fetchDevLinks from '../../services/api';
+import fetchDevLinks from '../../services/api/api';
 
 import {
   githubOAuthLogin,
